move df from prepped arrays to postproc_utils
